minor style fixes
