+++ Central-Server/API/internal/auth.py
@@ -1,4 +1,6 @@
 """This file will eventually contain all the functions necessary to determine who a user is and what they can do."""
-def authenticate_request(xhr) -> bool:
+from flask import Request
+
+def authenticate_request(xhr: Request) -> bool:
     """Determine if the request has MEMBER-status permission (user is a member of ISSUIUC)"""
     return True # TODO: actual checks
